+++ docs/CHANGELOG.md
@@ -17,6 +17,8 @@ _Latest build from master branch._
   ([#176](https://github.com/Tubeshade/Tubeshade/issues/176))
 - Use yt-dlp built-in logic for downloading the best thumbnail
   ([#238](https://github.com/Tubeshade/Tubeshade/pull/238))
+- Show rate/remaining estimate for more tasks
+  ([#88](https://github.com/Tubeshade/Tubeshade/issues/88))
 
 ### Fixed